pyproject.toml
[project]
name = "book-sharing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt==4.0.1",
    "fastapi>=0.115.12",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email,emil]>=2.11.5",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.2",
]
